+++ CHANGES.md
@@ -6,6 +6,7 @@
 - fix(client): Use HTTP/1.1 when ALPN is not supported by remote server.
 - fix(client): Preventing from memory leak when `h2` is selected by ALPN.
 - fix(client): Allow string-type content-length.
+- fix(client): Renamed `EVersion.AUTO` to `EVersion.ALPN`.
 - build(deps): removed all runtime dependencies.
 
 ## v1.0.5